Refactor nightscout_display.py to utilize PygameDisplay class and remove unused functions; update .gitignore to include CHAT*
